# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Order Fulfillment and Logistics Hub API
  version: 0.1.0
  description: >-
    Contract for BUSINESS administration, trusted internal intake, and isolated
    DemoScenario operations. JSON field names use snake_case; identifiers and
    NUMERIC(18,3) quantities are strings.
servers:
  - url: http://localhost:8080
    description: Local Docker Compose
...
security:
  - basicAuth: []
paths:
...
  /api/v1/import-batches/{batch_id}/source-return-exports:
    parameters:
      - $ref: '#/components/parameters/BatchId'
    get:
      tags: [Imports, Exports]
      operationId: listSourceReturnExports
      responses:
        '200':
          description: Zero or one final source-format return file; no interim file is exposed before all tracking facts are ready
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SourceReturnExport' }
        default: { $ref: '#/components/responses/ErrorResponse' }
  /api/v1/source-return-exports/{export_id}/file:
    parameters:
      - $ref: '#/components/parameters/ExportId'
    get:
      tags: [Exports]
      operationId: downloadSourceReturnFile
      parameters:
        - $ref: '#/components/parameters/Operator'
      responses:
        '200':
          description: Original-channel workbook or CSV
          headers:
            X-Request-Id: { $ref: '#/components/headers/RequestId' }
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        default: { $ref: '#/components/responses/ErrorResponse' }
...
  /api/v1/fulfillment-exports/{export_id}/tracking-imports:
    parameters:
      - $ref: '#/components/parameters/ExportId'
    post:
      tags: [Exports, Imports]
      operationId: uploadProviderTrackingFile
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/Operator'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                import_mode: { $ref: '#/components/schemas/ImportMode' }
                parent_import_batch_id: { $ref: '#/components/schemas/Identifier' }
      responses:
        '201':
          description: >-
            Whole provider batch accepted atomically and linked to this export. Replaying
            identical bytes for the same export returns the first complete result, including
            business_results and generated_source_return_export_ids, without a new source return.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TrackingImportBatch' }
        default: { $ref: '#/components/responses/ErrorResponse' }
  /api/v1/tracking-imports/{batch_id}:
    parameters:
      - $ref: '#/components/parameters/BatchId'
    get:
      tags: [Imports]
      operationId: getTrackingImport
      responses:
        '200':
          description: Provider result rows and source-return generation result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TrackingImportBatch' }
        default: { $ref: '#/components/responses/ErrorResponse' }
  /api/v1/shipments:
    get:
      tags: [Shipments]
      operationId: listShipments
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/DateFrom'
        - $ref: '#/components/parameters/DateTo'
        - name: provider_id
          in: query
          schema: { $ref: '#/components/schemas/Identifier' }
        - name: shipment_status
          in: query
          schema: { $ref: '#/components/schemas/ShipmentStatus' }
      responses:
        '200':
          description: Shipment page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ShipmentPage' }
        default: { $ref: '#/components/responses/ErrorResponse' }
...
  /api/v1/shipments/{shipment_id}/jd-tracking-backfill:
    parameters:
      - $ref: '#/components/parameters/ShipmentId'
    post:
      tags: [Shipments]
      operationId: backfillShipmentJdTracking
      description: >-
        Queries JD querySoOrder by the Shipment's stable erpDeliveryNo and, only after the
        persisted client mode, JD delivery reference, submitted warehouse, owner authority and cargo snapshots,
        plus the provider's currently configured pin credential,
        exact shipped quantities, and single carrier/waybill all match, atomically accepts one
        Tracking for the complete Shipment. The JD call is read-only and runs outside the database transaction.
        Replaying the same Idempotency-Key returns the original result without another JD query.
        Pending or partial results do not advance Shipment, OrderLine, or Order state; split or
        conflicting tracking opens/reuses MULTIPLE_TRACKINGS_FOR_OUTBOUND ReviewCase. Carrier
        ambiguity and JD terminal exceptions open their own blocking ReviewCases. The submitted
        owner is durable authority; the current pin is ephemeral and is never returned or audited.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/Operator'
      responses:
        '200':
          description: >-
            Poll result. QUERY_FAILED is a persisted, retryable business result; callers use a
            new idempotency key for a later read attempt. No raw JD response or receiver PII is returned.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ShipmentJdTrackingBackfillResult' }
        '409':
          description: >-
            The local shipment/outbound facts changed, a prior claim is still in progress, the
            persisted submission client mode differs from the current JD client, historical data
            lacks trustworthy submitted warehouse/owner/cargo snapshots, or the current provider
            pin is missing, or the terminal exception was already closed by an operator
            (JD_TRACKING_SUBMITTED_OWNER_MISSING / JD_TRACKING_PIN_CONFIG_MISSING /
            JD_TRACKING_TERMINAL_EXCEPTION_REVIEWED).
        default: { $ref: '#/components/responses/ErrorResponse' }
...
  /api/v1/shipments/{shipment_id}/jd-stock-check:
    parameters:
      - $ref: '#/components/parameters/ShipmentId'
    post:
      tags: [Shipments, Inventory]
      operationId: checkShipmentJdStock
      summary: 使用当前京东出库预览执行实时库存判定
      description: >-
        复用当前 Shipment 出库预览的 goodsNo 与精确京东件数，在事务外查询目标京东仓。
        缺行、重复行、无效数量、调用失败和库存不足均默认阻断；该结果不是库存预占，
        不创建京东出库单、不切换履约方，也不创建公司自营采购工单。相同幂等键重放
        首次结果，提交出库单前仍必须再次执行实时检查。若当前出库预览本身不可提交，
        请求以 409 前置条件冲突失败，不调用京东库存查询。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/Operator'
      responses:
        '200':
          description: 在可提交预览前提下的库存通过或可操作阻断结果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ShipmentJdStockCheckResult' }
        default: { $ref: '#/components/responses/ErrorResponse' }
...
  /api/v1/inventory/details:
    get:
      tags: [Inventory]
      operationId: getInventoryDetails
      summary: 读取单个履约方 SKU 的专业库存明细
      description: >-
        以履约方、SKU 和可选仓库为业务对象，返回主数据上下文、最新已落库库存观测与
        当前明确接入的专业库存查询能力。本接口只读取本地快照，不调用履约方接口；
        未观测时数量保持 null，不把缺失事实伪造为 0。京东低频原始查询仍由系统设置下的
        京东工具承载，能力中的 MOCK/REAL 标记描述当前客户端运行模式，查询结果不会改写
        权威库存事实。
      parameters:
        - name: provider_id
          in: query
          required: true
          schema: { $ref: '#/components/schemas/Identifier' }
        - name: sku_id
          in: query
          required: true
          schema: { $ref: '#/components/schemas/Identifier' }
        - name: warehouse_code
          in: query
          description: 可选目标观测仓库；无该仓观测时仍返回对象上下文和 NOT_OBSERVED。
          schema: { type: string, minLength: 1, maxLength: 128, pattern: '^\S+$' }
      responses:
        '200':
          description: 业务对象上下文、缓存观测、新鲜度与已接入能力
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InventoryDetailsResponse' }
        default: { $ref: '#/components/responses/ErrorResponse' }
...
  /api/v1/operational-alerts:
    get:
      tags: [Reviews]
      operationId: listOperationalAlerts
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - name: status
          in: query
          schema: { type: string, enum: [OPEN, ACKNOWLEDGED, RESOLVED] }
        - name: severity
          in: query
          schema: { type: string, enum: [YELLOW, RED] }
      responses:
        '200':
          description: Non-blocking operational alerts
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OperationalAlertPage' }
        default: { $ref: '#/components/responses/ErrorResponse' }
  /api/v1/operational-alerts/{alert_id}/acknowledge:
    parameters:
      - $ref: '#/components/parameters/AlertId'
    post:
      tags: [Reviews]
      operationId: acknowledgeOperationalAlert
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/Operator'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VersionedNoteCommand' }
      responses:
        '200':
          description: Alert acknowledged without changing business state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OperationalAlert' }
        '409':
          description: expected_version is stale or the alert is no longer OPEN
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        default: { $ref: '#/components/responses/ErrorResponse' }
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: Nginx authenticates management, internal, and demo operations; explicitly public connector callbacks use their own signatures.
  headers:
    RequestId:
      description: Request correlation identifier
      schema: { type: string }
  responses:
    ErrorResponse:
      description: Unified business or technical error
      headers:
        X-Request-Id: { $ref: '#/components/headers/RequestId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
...
    DateFrom:
      name: date_from
      in: query
      schema: { type: string, format: date }
    DateTo:
      name: date_to
      in: query
      schema: { type: string, format: date }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 255 }
    Operator:
      name: X-Operator
      in: header
      required: true
      description: Required at the backend seam; the public gateway overwrites this value with the authenticated principal.
      schema: { type: string, minLength: 1, maxLength: 128 }
...
    BatchId: { name: batch_id, in: path, required: true, schema: { $ref: '#/components/schemas/Identifier' } }
    ExportId: { name: export_id, in: path, required: true, schema: { $ref: '#/components/schemas/Identifier' } }
...
    ShipmentId: { name: shipment_id, in: path, required: true, schema: { $ref: '#/components/schemas/Identifier' } }
...
    AlertId: { name: alert_id, in: path, required: true, schema: { $ref: '#/components/schemas/Identifier' } }
...
  schemas:
    Identifier:
      type: string
      pattern: '^[1-9][0-9]*$'
      example: '42'
...
    ImportMode:
      type: string
      enum: [NEW, REVISION]
      default: NEW
...
    ShipmentStatus:
      type: string
      enum: [CREATED, SHIPPED, FAILED, DELIVERED]
...
    PageMeta:
      type: object
      required: [page, size, total_elements, total_pages]
      properties:
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1 }
        total_elements: { type: integer, format: int64, minimum: 0 }
        total_pages: { type: integer, minimum: 0 }
...
    ApiError:
      type: object
      required: [business_code, message, http_status, request_id, trace_id]
      properties:
        business_code: { type: string, maxLength: 64 }
        message: { type: string }
        http_status: { type: integer, minimum: 400, maximum: 599 }
        request_id: { type: string }
        trace_id: { type: string }
        field_errors:
          type: array
          items: { $ref: '#/components/schemas/FieldError' }
        details:
          type: object
          additionalProperties: true
...
    ImportBatch:
      type: object
      required: [id, batch_no, batch_type, import_mode, template_family, template_version, original_file_name, content_sha256, status, row_counts, received_at]
      properties:
        id: { $ref: '#/components/schemas/Identifier' }
        batch_no: { type: string }
        batch_type: { type: string, enum: [SOURCE_ORDER, PROVIDER_TRACKING] }
        import_mode: { $ref: '#/components/schemas/ImportMode' }
        parent_import_batch_id: { $ref: '#/components/schemas/Identifier' }
        revision_no: { type: integer, minimum: 1 }
        source_channel: { $ref: '#/components/schemas/SourceChannel' }
        fulfillment_provider_id: { $ref: '#/components/schemas/Identifier' }
        source_fulfillment_export_id: { $ref: '#/components/schemas/Identifier' }
        template_family: { type: string }
        template_version: { type: string }
        template_fingerprint: { type: string }
        original_file_name: { type: string }
        content_sha256: { type: string, pattern: '^[0-9a-f]{64}$' }
        status: { $ref: '#/components/schemas/ImportBatchStatus' }
        confirmed_at: { type: string, format: date-time, nullable: true }
        confirmed_by: { type: string, nullable: true }
        row_counts:
          type: object
          required: [total, accepted, need_review, rejected]
          properties:
            total: { type: integer, minimum: 0 }
            accepted: { type: integer, minimum: 0 }
            need_review: { type: integer, minimum: 0 }
            rejected: { type: integer, minimum: 0 }
        generated_fulfillment_export_ids:
          type: array
          items: { $ref: '#/components/schemas/Identifier' }
        generated_source_return_export_ids:
          type: array
          items: { $ref: '#/components/schemas/Identifier' }
        outbound_routing:
          type: object
          nullable: true
          description: >-
            仅首次确认响应携带（jd-real-sdk-switch 05）：京东 SDK 直连路由的建单发货批次
            （shipment_id 列表）；幂等重放与后续读取不出现。
          properties:
            jd_sdk_shipment_ids:
              type: array
              items: { type: string }
              description: SDK 直连路由的建单发货批次（shipment_id）。
        error_detail: { type: object, additionalProperties: true }
        received_at: { type: string, format: date-time }
        processed_at: { type: string, format: date-time, nullable: true }
...
    RawImportRow:
      type: object
      required: [id, sheet_name, sheet_index, row_index, raw_cells, status]
      properties:
        id: { $ref: '#/components/schemas/Identifier' }
        sheet_name: { type: string }
        sheet_index: { type: integer, minimum: 0 }
        row_index: { type: integer, minimum: 1 }
        raw_cells:
          oneOf:
            - { type: object, additionalProperties: true }
            - { type: array, items: {} }
        source_order_ref: { type: string }
        status: { $ref: '#/components/schemas/RawRowStatus' }
        error_code: { type: string }
        error_detail: { type: object, additionalProperties: true }
        order_id: { $ref: '#/components/schemas/Identifier' }
        order_line_id: { $ref: '#/components/schemas/Identifier' }
        parsed:
          type: object
          description: 渠道模板解析投影（白名单：receiver_name/receiver_phone/receiver_address/product_name/quantity/specification/source_sku_ref），供确认明细核对解析是否正确。
          additionalProperties: { type: string }
        sku_fulfillment:
          type: object
          nullable: true
          description: 来源 SKU 归属的履约方（内部 SKU 主数据）；无有效映射为 null。
          required: [provider_type, provider_name]
          properties:
            provider_type:
              type: string
              enum: [JD_WAREHOUSE, THIRD_PARTY]
            provider_name: { type: string }
            sku_specification:
              type: string
              description: 内部 SKU 主数据规格（来源文件未提供规格时的默认值）。
...
    Shipment:
      type: object
      required: [id, shipment_no, order_id, provider_id, outbound_order_no, shipment_sequence, shipment_status, receiver, items, created_at, updated_at]
      properties:
        id: { $ref: '#/components/schemas/Identifier' }
        shipment_no: { type: string }
        order_id: { $ref: '#/components/schemas/Identifier' }
        provider_id: { $ref: '#/components/schemas/Identifier' }
        outbound_order_no:
          type: string
          pattern: '^[0-9]{12}$'
          description: System-generated Shanghai yyyyMMdd plus four-digit daily atomic sequence.
        shipment_sequence: { type: integer, minimum: 1 }
        shipment_status: { $ref: '#/components/schemas/ShipmentStatus' }
        receiver: { $ref: '#/components/schemas/Receiver' }
        items:
          type: array
          items: { $ref: '#/components/schemas/ShipmentItem' }
        tracking: { $ref: '#/components/schemas/Tracking' }
        jd_outbound:
          allOf:
            - $ref: '#/components/schemas/ShipmentJdOutbound'
          nullable: true
          description: >-
            Shipment-level JD outbound integration record (at most one per Shipment).
            Exposes only diagnostic fields — never credentials, request payloads, or raw PII.
        shipped_at:
          type: string
          format: date-time
          nullable: true
          description: Actual shipment time when supplied; null means the provider did not provide one and is never replaced by receipt or confirmation time.
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    ShipmentPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Shipment' }
...
    ShipmentJdTrackingBackfillResult:
      type: object
      required: [shipment_id, erp_delivery_no, poll_status, retryable]
      properties:
        shipment_id: { $ref: '#/components/schemas/Identifier' }
        erp_delivery_no:
          type: string
          description: Stable merchant-side outbound reference used for querySoOrder.
        poll_status:
          type: string
          enum: [PENDING, PARTIAL, TRACKED, CONFLICT, QUERY_FAILED]
        retryable:
          type: boolean
          description: True only for a failed read result that may be retried with a new idempotency key.
        tracking_number:
          type: string
          nullable: true
          description: Present only after a unique full result was accepted or idempotently replayed.
        review_case_id:
          allOf: [{ $ref: '#/components/schemas/Identifier' }]
          nullable: true
          description: >-
            Present when split/conflicting tracking, Carrier mapping, or a JD terminal exception
            requires human review.
        business_code:
          type: string
          nullable: true
          description: >-
            Sanitized stable code for QUERY_FAILED or CONFLICT, including
            JD_TRACKING_CARRIER_MAPPING_REQUIRED and JD_TRACKING_TERMINAL_EXCEPTION.
...
    TrackingImportBatch:
      allOf:
        - $ref: '#/components/schemas/ImportBatch'
        - type: object
          required: [source_fulfillment_export_id]
          properties:
            source_fulfillment_export_id: { $ref: '#/components/schemas/Identifier' }
            business_results:
              type: object
              properties:
                shipped: { type: integer, minimum: 0 }
                partial: { type: integer, minimum: 0 }
                failed: { type: integer, minimum: 0 }
            rows:
              type: array
              items: { $ref: '#/components/schemas/RawImportRow' }
    SourceReturnExport:
      type: object
      required: [id, import_batch_id, version_no, is_final, template_version, tracking_cutoff_at, file_sha256, generated_at]
      properties:
        id: { $ref: '#/components/schemas/Identifier' }
        import_batch_id: { $ref: '#/components/schemas/Identifier' }
        version_no: { type: integer, minimum: 1 }
        is_final: { type: boolean }
        template_version: { type: string }
        tracking_cutoff_at: { type: string, format: date-time }
        file_sha256: { type: string, pattern: '^[0-9a-f]{64}$' }
        generated_at: { type: string, format: date-time }
...
    VersionedNoteCommand:
      type: object
      required: [expected_version, note]
      properties:
        expected_version: { type: integer, minimum: 0 }
        note: { type: string, minLength: 1, maxLength: 1000 }
    OperationalAlert:
      type: object
      required: [id, alert_no, alert_type, severity, status, message, detail, version, created_at]
      properties:
        id: { $ref: '#/components/schemas/Identifier' }
        alert_no: { type: string }
        alert_type: { type: string }
        severity: { type: string, enum: [YELLOW, RED] }
        status: { type: string, enum: [OPEN, ACKNOWLEDGED, RESOLVED] }
        order_id: { $ref: '#/components/schemas/Identifier' }
        order_line_id: { $ref: '#/components/schemas/Identifier' }
        fulfillment_id: { $ref: '#/components/schemas/Identifier' }
        shipment_id: { $ref: '#/components/schemas/Identifier' }
        message: { type: string }
        detail: { type: object, additionalProperties: true }
        acknowledged_by: { type: string, nullable: true }
        acknowledged_at: { type: string, format: date-time, nullable: true }
        resolved_at: { type: string, format: date-time, nullable: true }
        version: { type: integer, minimum: 0 }
        created_at: { type: string, format: date-time }
    OperationalAlertPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/OperationalAlert' }
...
    ShipmentJdStockBlocker:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - JD_SKU_MAPPING_GATE_BLOCKED
            - JD_STOCK_QUERY_FAILED
            - JD_STOCK_RESPONSE_INVALID
            - JD_STOCK_RESPONSE_AMBIGUOUS
            - JD_STOCK_TARGET_WAREHOUSE_NOT_OBSERVED
            - JD_STOCK_INSUFFICIENT
        message: { type: string }
    ShipmentJdStockObservation:
      type: object
      additionalProperties: false
      required: [sku_id, goods_no, warehouse_code, required_quantity, quantity_unit, observation_status]
      properties:
        sku_id: { $ref: '#/components/schemas/Identifier' }
        goods_no: { type: string, minLength: 1 }
        warehouse_code: { type: string, minLength: 1 }
        required_quantity:
          type: string
          pattern: '^[1-9][0-9]*$'
          description: 当前出库预览所需的精确京东件数。
        quantity_unit: { type: string, enum: [JD_PIECE] }
        observation_status: { type: string, enum: [OBSERVED, OBSERVED_ZERO, NOT_OBSERVED] }
        stock_quantity: { $ref: '#/components/schemas/DecimalQuantity' }
        usable_quantity: { $ref: '#/components/schemas/DecimalQuantity' }
    ShipmentJdStockReviewCase:
      type: object
      additionalProperties: false
      required: [id, reason_code, status]
      properties:
        id: { $ref: '#/components/schemas/Identifier' }
        reason_code: { type: string, enum: [JD_STOCK_BLOCKED] }
        status: { type: string, enum: [OPEN] }
    ShipmentJdStockCheckResult:
      type: object
      additionalProperties: false
      required:
        - shipment_id
        - shipment_version
        - preview_hash
        - target_warehouse_code
        - stock_status
        - observation_status
        - observed_at
        - not_reserved
        - blockers
        - items
      properties:
        shipment_id: { $ref: '#/components/schemas/Identifier' }
        shipment_version: { type: integer, format: int64, minimum: 0 }
        preview_hash: { type: string, pattern: '^[0-9a-f]{64}$' }
        target_warehouse_code:
          type: string
          minLength: 1
        stock_status: { type: string, enum: [PASSED, BLOCKED] }
        observation_status: { type: string, enum: [OBSERVED, OBSERVED_ZERO, NOT_OBSERVED] }
        observed_at: { type: string, format: date-time }
        not_reserved:
          type: boolean
          enum: [true]
          description: 实时查询结果是建议性事实，不代表京东已预占库存。
        blockers:
          type: array
          items: { $ref: '#/components/schemas/ShipmentJdStockBlocker' }
        items:
          type: array
          items: { $ref: '#/components/schemas/ShipmentJdStockObservation' }
        review_case: { $ref: '#/components/schemas/ShipmentJdStockReviewCase' }
...
    InventoryDetailsResponse:
      type: object
      additionalProperties: false
      required: [context, observation, query_time, freshness_policy, capabilities]
      properties:
        context: { $ref: '#/components/schemas/InventoryDetailContext' }
        observation: { $ref: '#/components/schemas/InventoryDetailObservation' }
        query_time: { type: string, format: date-time }
        freshness_policy:
          type: string
          pattern: '^P(?!$).+$'
          description: ISO-8601 Duration；当前默认 PT15M。
        capabilities:
          type: array
          items: { $ref: '#/components/schemas/InventoryDetailCapability' }
...